問題[https://github.com/Hexirp/monadiclasses/issues/1] のための修正
