+++ mini/Structure/Monadiclasses/Minimal.hs
@@ -1,6 +1,8 @@
 {-# LANGUAGE NoImplicitPrelude #-}
 
 module Structure.Monadiclasses.Minimal (
-    module Structure.Monadiclasses.Minimal.Functor
+    module Structure.Monadiclasses.Minimal.Functor,
+    module Structure.Monadiclasses.Minimal.Comonad
 ) where
     import Structure.Monadiclasses.Minimal.Functor
+    import Structure.Monadiclasses.Minimal.Comonad
